# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "First_proj",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "description": "Returns order data",
                "operationId": "getOrderByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User's id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 28
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "total_price": {
                                            "type": "string",
                                            "example": "700"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-08-15T08:55:40.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-08-15T08:55:40.000000Z"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/orders/delete/{id}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete the Order",
                "description": "Returns response",
                "operationId": "deleteOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order's id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order was removed"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}